+++ R/utilities.R
@@ -1,42 +1,26 @@
+get_dose_env <- function() {
+    if (!exists(".DOSEEnv")) {
+        .initial()
+    }
+    get(".DOSEEnv")
+}
+
 .initial <- function() {
     pos <- 1
     envir <- as.environment(pos)
-    assign(".DOSEEnv", new.env(), envir = envir)
-    .DOSEEnv <- get(".DOSEEnv", envir = envir)
-
-    tryCatch(utils::data(list="dotbl",
-                         package="DOSE"))
-    dotbl <- get("dotbl")
-    assign("dotbl", dotbl, envir = .DOSEEnv)
-    rm(dotbl, envir = .GlobalEnv)
-
-    tryCatch(utils::data(list="mpotbl",
-                         package="DOSE"))
-    mpotbl <- get("mpotbl")
-    assign("mpotbl", mpotbl, envir = .DOSEEnv)
-    rm(mpotbl, envir = .GlobalEnv)
-
-    tryCatch(utils::data(list="hpotbl",
-                         package="DOSE"))
-    hpotbl <- get("hpotbl")
-    assign("hpotbl", hpotbl, envir = .DOSEEnv)
-    rm(hpotbl, envir = .GlobalEnv)
-
-    tryCatch(utils::data(list="DOIC",
-                         package="DOSE"))
-    DOIC <- get("DOIC")
-    assign("DOIC", DOIC, envir = .DOSEEnv)
-    rm(DOIC, envir = .GlobalEnv)
-    
+    assign(".DOSEEnv", new.env(), envir = envir) 
 }
 
+# https://github.com/YuLab-SMU/ReactomePA/issues/43
 check_gene_id <- function(geneList, geneSets) {
     if (all(!names(geneList) %in% unique(unlist(geneSets)))) {
         sg <- unlist(geneSets[1:10])
         sg <- sample(sg, min(length(sg), 6))
         message("--> Expected input gene ID: ", paste0(sg, collapse=','))
-        stop("--> No gene can be mapped....")
+        message("--> No gene can be mapped....")
+        return(FALSE)
     }
+    return(TRUE)
 }
 
 
@@ -65,64 +49,6 @@ calculate_qvalue <- function(pvals) {
 }
 
 
-prepare_relation_df <- function() {
-    gtb <- toTable(HDOTERM)
-    gtb <- gtb[,1, drop=FALSE]
-    gtb <- unique(gtb)
-
-    id <- gtb$do_id
-    pid <- mget(id, HDOPARENTS)
-    cid <- rep(names(pid), times=sapply(pid, length))
-
-    ptb <- data.frame(id=cid,
-                      relationship = 'other',
-                      parent = unlist(pid),
-                      Ontology = "DO",
-                      stringsAsFactors = FALSE)
-
-    dotbl <- merge(gtb, ptb, by.x="doid", by.y="id")
-    save(dotbl, file="dotbl.rda", compress="xz")
-    invisible(dotbl)
-    # mpotbl
-    gtb <- toTable(MPO.db::MPOTERM)
-    gtb <- gtb[,1, drop=FALSE]
-    gtb <- unique(gtb)
-    id <- gtb$do_id
-    pid <- mget(id, MPO.db::MPOPARENTS)
-    cid <- rep(names(pid), times=sapply(pid, length))
-
-    ptb <- data.frame(id=cid,
-                      relationship = 'other',
-                      parent = unlist(pid),
-                      Ontology = "MPO",
-                      stringsAsFactors = FALSE)
-    
-    mpotbl <- merge(gtb, ptb, by.x="mpid", by.y="id")
-    save(mpotbl, file="mpotbl.rda", compress="xz")
-    invisible(mpotbl)
-
-    # hpotbl
-    gtb <- toTable(HPO.db::HPOTERM)
-    gtb <- gtb[,1, drop=FALSE]
-    gtb <- unique(gtb)
-    id <- gtb$do_id
-    pid <- mget(id, HPO.db::HPOPARENTS)
-    cid <- rep(names(pid), times=sapply(pid, length))
-
-    ptb <- data.frame(id=cid,
-                      relationship = 'other',
-                      parent = unlist(pid),
-                      Ontology = "HPO",
-                      stringsAsFactors = FALSE)
-    
-    hpotbl <- merge(gtb, ptb, by.x="hpoid", by.y="id")
-    save(hpotbl, file="hpotbl.rda", compress="xz")
-    invisible(hpotbl)
-}
-
-
-
-
 calculate_qvalue <- function(pvals) {
     if (length(pvals) == 0)
         return(numeric(0))
@@ -142,36 +68,14 @@ calculate_qvalue <- function(pvals) {
 ##'
 ##'
 ##' @title compute information content
-##' @param ont one of "DO" and "MPO"
+##' @param ont one of "DO", "HPO" and "MPO"
 ##' @return NULL
 ##' @importMethodsFrom AnnotationDbi toTable
 ##' @author Guangchuang Yu \url{https://yulab-smu.top}
-computeIC <- function(ont="DO"){
-    if (!exists(".DOSEEnv")) {
-        .initial()
-    }
-    ont <- match.arg(ont, c("DO", "MPO", "HPO"))
-    DOSEEnv <- get(".DOSEEnv", envir = .GlobalEnv)
-    if (ont == "DO") {
-        if (!exists("DO2EG", envir=DOSEEnv)) {
-            tryCatch(utils::data(list="DO2EG", package="DOSE"))
-            assign("DO2EG", DO2EG, envir = DOSEEnv)
-            DO2EG <- get("DO2EG")
-            rm(DO2EG, envir = .GlobalEnv)
-        }
-        DO2EG <- get("DO2EG", envir = DOSEEnv)
-        Offsprings <- AnnotationDbi::as.list(HDO.db::HDOOFFSPRING)
-    } else if (ont == "MPO") {
-        eg.do <- toTable(MPO.db::MPOMPMGI)[, c(2,1)]
-        colnames(eg.do) <- c("eg", "doid")
-        DO2EG <- with(eg.do, split(as.character(eg), as.character(doid)))
-        Offsprings <- AnnotationDbi::as.list(MPO.db::MPOOFFSPRING)
-    } else if (ont == "HPO") {
-        eg.do <- toTable(HPO.db::HPOGENE)[, c(2,1)]
-        colnames(eg.do) <- c("eg", "doid")
-        DO2EG <- with(eg.do, split(as.character(eg), as.character(doid)))
-        Offsprings <- AnnotationDbi::as.list(HPO.db::HPOOFFSPRING)
-    }
+computeIC <- function(ont="HDO"){
+    DO2EG <- get_ont2gene(ont)
+    Offsprings <- GOSemSim:::getOffsprings(ont)
+    
     docount <- unlist(lapply(DO2EG, length))
     doids <- names(docount) 
     
@@ -197,34 +101,10 @@ computeIC <- function(ont="DO"){
 ##' @importMethodsFrom AnnotationDbi exists
 ##' @export
 ##' @author Guangchuang Yu \url{https://yulab-smu.top}
-gene2DO <- function(gene, organism = "hsa", ont = "DO") {
+gene2DO <- function(gene, organism = "hsa", ont = "HDO") {
     gene <- as.character(gene)
-    if (organism == "hsa") {
-        if(!exists(".DOSEEnv")) .initial()
-        .DOSEEnv <- get(".DOSEEnv", envir=.GlobalEnv)
-        if (!exists("EG2DO", envir = .DOSEEnv)) {
-            tryCatch(utils::data(list="EG2DO", package="DOSE"))
-            EG2DO <- get("EG2DO")
-            assign("EG2DO", EG2DO, envir=.DOSEEnv)
-            rm(EG2DO, envir=.GlobalEnv)
-        }
-        EG2DO <- get("EG2DO", envir=.DOSEEnv)
-    } else {
-        if (ont == "DO") {
-            eg.do <- toTable(MPO.db::MPOMGIDO)
-            colnames(eg.do) <- c("eg", "doid")
-            MPOTERMs <- names(as.list(HDOANCESTOR))             
-        } else {
-            eg.do <- toTable(MPO.db::MPOMPMGI)[, c(2,1)]
-            colnames(eg.do) <- c("eg", "doid")
-            MPOTERMs <- names(as.list(MPO.db::MPOANCESTOR))
-        }
-        EG2DO <- with(eg.do, split(as.character(doid), as.character(eg)))
-        EG2DO <- lapply(EG2DO, function(i) unique(i[ i %in% MPOTERMs ]))
-        i <- unlist(lapply(EG2DO, function(i) length(i) != 0))
-        EG2DO <- EG2DO[i]   
-    
-    }
+
+    EG2DO <- get_gene2ont(ont)
 
     DO <- EG2DO[[gene]]
     DO <- unlist(DO)
@@ -242,133 +122,57 @@ gene2DO <- function(gene, organism = "hsa", ont = "DO") {
 }
 
 process_tcss <- getFromNamespace("process_tcss", "GOSemSim")
-##' @importClassesFrom GOSemSim GOSemSimDATA
-mpodata <- function(processTCSS = FALSE) {
-    if (!exists(".DOSEEnv")) .initial()
-    DOIC <- new("GOSemSimDATA",
-                  ont = "MPO",
-                  IC = computeIC(ont = "MPO"))    
-    if (processTCSS) {
-        IC <- DOIC@IC
-        DOIC@tcssdata <- process_tcss(ont = "MPO", IC = IC, cutoff = NULL)
-    }
-    DOIC
-}
 
 ##' @importClassesFrom GOSemSim GOSemSimDATA
-hpodata <- function(processTCSS = FALSE) {
-    if (!exists(".DOSEEnv")) .initial()
-    DOIC <- new("GOSemSimDATA",
-                  ont = "HPO",
-                  IC = computeIC(ont = "HPO"))    
+semdata <- function(processTCSS = FALSE, ont = "HDO") {
+    IC <- new("GOSemSimDATA",
+                ont = ont,
+                IC = computeIC(ont = ont))
+
     if (processTCSS) {
-        IC <- DOIC@IC
-        DOIC@tcssdata <- process_tcss(ont = "HPO", IC = IC, cutoff = NULL)
+        IC <- IC@IC
+        IC@tcssdata <- process_tcss(ont = ont, IC = IC, cutoff = NULL)
     }
-    DOIC
+
+    IC
 }
 
+semdata2 <- memoise::memoise(semdata)
 
-##' @importClassesFrom GOSemSim GOSemSimDATA
-dodata <- function(processTCSS = FALSE) {
-    if (!exists(".DOSEEnv")) .initial()
-    .DOSEEnv <- get(".DOSEEnv", envir=.GlobalEnv)
-    DOIC <- get("DOIC", envir=.DOSEEnv)
-    if (processTCSS) {
-        IC <- DOIC@IC
-        DOIC@tcssdata <- process_tcss(ont = "DO", IC = IC, cutoff = NULL)
+
+get_ont2gene <- function(ontology, output = "list") {
+    gene2ont <- get_gene2ont(ontology, output = "data.frame")
+    if (output == "data.frame") {
+        return(gene2ont[, 2:1])
     }
-    DOIC
+
+    split(as.character(gene2ont[,1]), as.character(gene2ont[,2]))
 }
 
-build_dodata <- function() {
-    
-    DOIC <- new("GOSemSimDATA",
-                  ont = "DO",
-                  IC = computeIC())
-    save(DOIC, file="DOIC.rda", compress="xz")
+get_gene2ont <- function(ontology, output = "list") {
+    ont2gene <- GOSemSim:::get_onto_data(ontology, table = "ont2gene", output = 'data.frame')
+    anc <- GOSemSim:::getAncestors(ontology)
+    idx <- ont2gene[,1] %in% names(anc)
+    ont2gene <- unique(ont2gene[idx, ])
+
+    if (output == "data.frame") {
+        return(ont2gene[, 2:1])
+    }
+
+    split(as.character(ont2gene[,1]), as.character(ont2gene[,2]))
 }
 
-
-##' rebuilding entrez and DO mapping datasets
-##'
-##'
-##' @title rebuiding annotation data
-##' @param file do_rif.human.txt
-##' @return NULL
-##' @importFrom utils read.delim
-##' @author Guangchuang Yu \url{https://yulab-smu.top}
-rebuildAnnoData <- function(file) {
-    ##
-    ## do_rif.human.txt was downloaded from
-    ## http://projects.bioinformatics.northwestern.edu/do_rif/
-    ##
-
-    ## do.rif <- read.delim2(file, sep="\t", stringsAsFactors=F, header=F)
-    ## eg.do <- do.rif[,c(1,5)]
-
-    ## new file
-    ## IDMappings.txt from
-    ## http://doa.nubic.northwestern.edu/pages/download.php
-    domapping <- read.delim(file, stringsAsFactors=F)
-    eg.do <- domapping[,c(2,1)]
-    colnames(eg.do) <- c("eg", "doid")
-    eg.do$doid <- paste("DOID:", eg.do$doid, sep="")
-
-    rebuildAnnoData.internal(eg.do)
+get_gene2allont <- function(ontology, output = "list") {
+    GOSemSim:::get_onto_data(ontology, table = "gene2allont", output = output)
 }
 
-##' @importFrom HDO.db HDOANCESTOR
-##' @importFrom HDO.db HDOTERM
-##' @importMethodsFrom AnnotationDbi mget
-rebuildAnnoData.internal <- function(eg.do) {
+get_ont2allgene <- function(ontology, output = "list") {
+    gene2allont <- GOSemSim:::get_onto_data(ontology, table = "gene2allont", output = "data.frame")
+    if (output == "data.frame") {
+        return(gene2allont[, 2:1])
+    }
 
-    eg <- doid <- NULL # to satisfy codetools
-
-    DO2EG <- with(eg.do, split(as.character(eg), as.character(doid)))
-    ## DO2EG <- dlply(eg.do, .(doid), .fun=function(i) i$eg)
-    # doids <- toTable(HDOTERM)
-    # HDOTERMs <- doids$do_id
-    HDOTERMs <- names(as.list(HDOANCESTOR))
-    idx <- names(DO2EG) %in% HDOTERMs
-    DO2EG <- DO2EG[idx]
-    DO2EG <- lapply(DO2EG, function(i) unique(i))
-    save(DO2EG, file="DO2EG.rda", compress="xz")
-
-    EG2DO <- with(eg.do, split(as.character(doid), as.character(eg)))
-    ## EG2DO <- dlply(eg.do, .(eg), .fun=function(i) i$doid)
-    EG2DO <- lapply(EG2DO, function(i) unique(i[ i %in% HDOTERMs ]))
-
-    i <- unlist(lapply(EG2DO, function(i) length(i) != 0))
-    EG2DO <- EG2DO[i]
-    save(EG2DO, file="EG2DO.rda", compress="xz")
-
-    EG2ALLDO <- lapply(EG2DO,
-                       function(i) {
-                           ans <- unlist(mget(i, HDOANCESTOR))
-                           ans <- ans[ !is.na(ans) ]
-                           ans <- c(i, ans)
-                           ans <- unique(ans)
-                           return(ans)
-                       })
-    save(EG2ALLDO, file="EG2ALLDO.rda", compress="xz")
-
-    len <- lapply(EG2ALLDO,length)
-    EG2ALLDO.df <- data.frame(EG=rep(names(EG2ALLDO), times=len),
-                              DO=unlist(EG2ALLDO))
-    DO <- NULL ## satisfy code tools
-    ## DO2ALLEG <- dlply(EG2ALLDO.df, .(DO), function(i) as.character(i$EG))
-    DO2ALLEG <- with(EG2ALLDO.df, split(as.character(EG), as.character(DO)))
-    DO2ALLEG <- lapply(DO2ALLEG, unique)
-    save(DO2ALLEG, file="DO2ALLEG.rda", compress="xz")
-
-
-    ## tryCatch(utils::data(list="DOSEEnv", package="DOSE"))
-    ## assign("DO2ALLEG", DO2ALLEG, envir=.DOSEEnv)
-    ## assign("EG2ALLDO", EG2ALLDO, envir=.DOSEEnv)
-    ## assign("EG2DO", EG2DO, envir=.DOSEEnv)
-    ## assign("DO2EG", DO2EG, envir=.DOSEEnv)
-    ## save(.DOSEEnv, file="DOSEEnv.rda", compress="xz")
+    split(as.character(gene2allont[,1]), as.character(gene2allont[,2]))
 }
 
 ## ##' get all entrezgene ID of a specific organism